Refactor settings page into modular sections
